Add files via upload
In these folders are the scripts, scenes, and assets for each version of the app. Due to the large file size of a Unity editable I was unable to upload the full files. If you want to edit it or make changes to it, you will have to put these scripts and scenes in a new Unity Project then reassign the proper variables in the scripts and put the scenes in the correct build order in the Build setting. You may also have to change the aspect ratio of the Game tab to be 16:9.
All of these versions have errors when building due to a built in Unity library. This is fixed in future versions. Also, this app needs some scale work done on it because when it is built and tested in an actual android environment it is way to small on the screen. Again, this issue is also fixed in future versions.